+++ react-app/src/App.js
@@ -1,26 +1,16 @@
 import logo from './logo.svg';
 import './App.css';
-import {BrowserRouter} from 'react-router-dom'
+import {BrowserRouter, Route, Routes} from 'react-router-dom'
+import { Base } from './pages/CalendarView';
+import { EmptyPlaceholder } from './pages/ViewTasks';
 
-function App() {
+function App(props) {
   return (
     <BrowserRouter>
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <p>
-            Edit <code>src/App.js</code> and save to reload.
-          </p>
-          <a
-            className="App-link"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-          </a>
-        </header>
-      </div>
+    <Routes>
+      <Route path="/calendar" element={<Base data={ props.data } />}></Route>
+      <Route path="/taskView" element={<EmptyPlaceholder data={props.data} />}></Route>
+    </Routes>
     </BrowserRouter>
   );
 }
