+++ react-app/src/pages/ViewTasks.jsx
@@ -1,5 +1,9 @@
-// import {Alert} from "react-bootstrap"
+import {Alert} from "react-bootstrap"
 
-// function EmptyPlaceholder() {
-//     return <Alert></Alert>
-// }
+function EmptyPlaceholder() {
+    return <Alert>Hi Cymbre</Alert>
+}
+
+export {
+    EmptyPlaceholder
+}
